Interpreter,ParseArgs: make --inferred-type print out the expression to be evaluated as well as the inferred type
The reason is that I discovered I had been led on a wild-hare chase for hours because I was typing in what I thought were valid expressions, but as it turns out, the shell was mangling them hopelessly! Arrgh!
So hopefully by enabling printing out of the expression as Mueval sees it, this'll happen less often, or at least be more diagnosable.

darcs-hash:20080726160651-f7719-98d3bdd73c1e29094a524891d25318f3595f35ea